+++ renderer/effects/events.js
@@ -1,0 +1,45 @@
+import queryString from 'query-string';
+import loadData from '../utils/data/load';
+import { API_EVENTS } from '../utils/data/endpoints';
+
+const loadEvents = (scope, dispatchEvents, { token }) => {
+  const defaults = { limit: 30 };
+  const query = Object.assign(defaults, {});
+
+  if (!scope.isCurrentUser) {
+    query.teamId = scope.id;
+  }
+
+  if (query.types) {
+    query.types = Array.from(query.types).join(',');
+  }
+
+  const params = queryString.stringify(query);
+
+  loadData(`${API_EVENTS}?${params}`, token).then(data => {
+    if (!data || !data.events) {
+      console.error(`Failed to load events: ${data}`);
+      return;
+    }
+
+    dispatchEvents({
+      scope: scope.id,
+      events: data.events
+    });
+  });
+};
+
+export default (scopes, active, events, dispatchEvents, config) => {
+  const scopesWithoutActive = scopes.filter(scope => {
+    return scope.id !== active.id;
+  });
+
+  // Bring the currently active scope to the front.
+  const orderedScopes = [active, ...scopesWithoutActive];
+
+  Promise.all(
+    orderedScopes.map(scope => {
+      return loadEvents(scope, dispatchEvents, config);
+    })
+  );
+};
